bracketplanner/requirements.txt
matplotlib==3.7.1
numpy==1.24.1
Pillow==9.5.0
scipy==1.10.1
warp==1.0.4
